query_service/requirements-query.txt
numpy<2.0
nltk
fastapi==0.110.0
uvicorn==0.27.0
httpx==0.26.0
pydantic==2.5.3
huggingface-hub==0.17.3
sentence-transformers==2.2.2
qdrant-client==1.6.4
pyyaml==6.0